+++ CHANGELOG.md
@@ -38,6 +38,7 @@
 - Added `quantiles` for inclusive interpolated cut points.
 - Added `sample_variance` for corrected one-pass variance.
 - Added `standard_deviation` for sample standard deviation.
+- Added `population_standard_deviation` for population standard deviation.
 
 ## [0.7.0] 2026-08-02
 ### Added